refactor: consolidate version packages
- Create shared/version package with standardized variable names
- Add Info() and Full() helper functions
- Update cfl and jtk to use shared version package
- Standardize on BuildDate (was Date in cfl)
- Update ldflags in Makefiles, .goreleaser.yml, and snapcraft.yaml
- Remove tool-specific version packages

Both tools now use consistent version formatting.

Closes #24
