+++ tools/cfl/Makefile
@@ -5,9 +5,9 @@ VERSION := $(shell git describe --tags --always --dirty 2>/dev/null || echo "dev
 COMMIT := $(shell git rev-parse --short HEAD 2>/dev/null || echo "unknown")
 DATE := $(shell date -u '+%Y-%m-%dT%H:%M:%SZ')
 LDFLAGS := -ldflags "-s -w \
-	-X github.com/open-cli-collective/confluence-cli/internal/version.Version=$(VERSION) \
-	-X github.com/open-cli-collective/confluence-cli/internal/version.Commit=$(COMMIT) \
-	-X github.com/open-cli-collective/confluence-cli/internal/version.Date=$(DATE)"
+	-X github.com/open-cli-collective/atlassian-go/version.Version=$(VERSION) \
+	-X github.com/open-cli-collective/atlassian-go/version.Commit=$(COMMIT) \
+	-X github.com/open-cli-collective/atlassian-go/version.BuildDate=$(DATE)"
 
 # Build the binary
 build:
