--- tools/cfl/internal/version/version.go
@@ -1,9 +1,0 @@
-// Package version provides build-time version information.
-package version
-
-// These variables are set at build time via ldflags.
-var (
-	Version = "dev"
-	Commit  = "unknown"
-	Date    = "unknown"
-)
