+++ tools/jtk/.goreleaser.yml
@@ -22,9 +22,9 @@ builds:
       - arm64
     ldflags:
       - -s -w
-      - -X github.com/open-cli-collective/jira-ticket-cli/internal/version.Version={{.Version}}
-      - -X github.com/open-cli-collective/jira-ticket-cli/internal/version.Commit={{.Commit}}
-      - -X github.com/open-cli-collective/jira-ticket-cli/internal/version.BuildDate={{.Date}}
+      - -X github.com/open-cli-collective/atlassian-go/version.Version={{.Version}}
+      - -X github.com/open-cli-collective/atlassian-go/version.Commit={{.Commit}}
+      - -X github.com/open-cli-collective/atlassian-go/version.BuildDate={{.Date}}
 
 archives:
   - id: default
